# V 1.0.0.2.3 沈阳分区赛 1. 采用地图整体模式，即无论在红方还是蓝方都采用同一张地图，同一个坐标系 2. 更改了发弹逻辑，移动到最后一个目标点后才允许发弹
# 请注意，当前这一版代码从无力到遥控器是正常的，但是从右下到遥控器会小陀螺正反狂转，我推测是一开始地图的vx方向反了，vx和vy反了，改代码时改出来的问题，现在雷达要到了，准备还原成雷达能用的状态

+++ PYRo/Component/Referee/protocol.h
@@ -226,7 +226,7 @@ struct robot_pos_t
     float x;
     float y;
     float angle;
-    uint32_t reserved;
+    float reserved;
 };
 
 // 0x0204
